factory-ai-vision/EdgeSolution/modules/WebModule/backend_v2/requirements/local.txt
-r ./base.txt
# Testing
# ---------------------------------------------------------
mypy==0.770
django-stubs==1.5.0
pytest==5.4.3
pytest-sugar==0.9.4
coverage==5.2.1

# Code quality
# ---------------------------------------------------------
isort==4.3.21
yapf==0.28.0
pylint==2.5.3
pylint-django==2.2.0

# Django
# ---------------------------------------------------------
factory-boy==2.12.0
django-coverage-plugin==1.8.0
pytest-django==3.9.0

# IPython
# ---------------------------------------------------------
IPython
